chore: factor out reportingRequirement form_change CRUD functions
